implement per-bucket history view
This commit includes a refactor of history functionality to support
bucket/bucket groups and rip out the existing history view code in
favor of using the same histogram system as realtime notices. This
makes history feature complete with the realtime notices view.

Note this also adds a new buckets field on the HistoryRecord, and
a corresponding (buckets, id) index. Existing history data should
backfill this field to maintain performance with this version.